+++ frontend/src/components/buttons/ButtonDelete.vue
