+++ frontend/src/components/camp/CollaboratorListItem.vue
@@ -25,7 +25,7 @@
 
 <script>
 import ApiSelect from '@/components/form/ApiSelect'
-import ButtonDelete from '@/components/base/ButtonDelete'
+import ButtonDelete from '@/components/buttons/ButtonDelete'
 
 export default {
   name: 'CollaboratorListItem',
